pyproject.toml
# Copyright (c) 2026 webAI, Inc.

[build-system]
requires = ["setuptools>=61.0,<78", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yolo-mlx"
version = "0.2.0"
description = "YOLO26 implementation using Apple MLX framework"
readme = "README.md"
license = "AGPL-3.0-only"
authors = [
    {name = "webAI Team"}
]
keywords = ["yolo", "object-detection", "object-tracking", "multi-object-tracking", "mlx", "apple-silicon", "deep-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
requires-python = ">=3.10"
dependencies = [
    "mlx>=0.30.3",
    "numpy>=2.0.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
tracking = [
    "opencv-python>=4.8.0",
    "lap>=0.5.0",
    "scipy>=1.11.0",
]
convert = [
    "torch>=2.0.0",
    "ultralytics>=8.0.0",
    "safetensors>=0.4.0",
]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
    "pre-commit>=3.7.0",
]

[project.urls]
Homepage = "https://github.com/thewebAI/yolo-mlx"
Documentation = "https://github.com/thewebAI/yolo-mlx#readme"
Repository = "https://github.com/thewebAI/yolo-mlx"
Changelog = "https://github.com/thewebAI/yolo-mlx/blob/main/CHANGELOG.md"

[project.scripts]
yolo-mlx = "yolo26mlx.cli:main"
yolomlx = "yolo26mlx.cli:main"
yolo26 = "yolo26mlx.cli:main"
yolo26mlx = "yolo26mlx.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
yolo26mlx = ["cfg/**/*.yaml"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310"]
